refactor: add maven structure
