+++ .gitignore
@@ -1,2 +1,4 @@
 # Project exclude paths
-/out/
+target
+.idea
+futebol-game.iml
