+++ .idea/modules.xml
@@ -2,7 +2,7 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
-      <module fileurl="file://$PROJECT_DIR$/FutebolGame.iml" filepath="$PROJECT_DIR$/FutebolGame.iml" />
+      <module fileurl="file://$PROJECT_DIR$/futebol-game.iml" filepath="$PROJECT_DIR$/futebol-game.iml" />
     </modules>
   </component>
 </project>
