+++ src/main/java/br/univali/kob/eng/exception/GameException.java
@@ -1,4 +1,4 @@
-package br.univali.eng.karoline.src.exception;
+package br.univali.kob.eng.exception;
 
 public class GameException extends Exception {
     public GameException(String exceptionName, String exceptionDescription){
